+++ src/battleship/window.c
@@ -16,6 +16,6 @@ void Window_Init(uint pos_x, uint pos_y, uint size_w, uint size_h)
 #ifdef _WIN32
    // https://batchloaf.wordpress.com/2012/04/17/how-a-console-application-can-set-the-size-of-its-window/
    HWND wh = GetConsoleWindow();
-   MoveWindow(wh, pos_x, pos_y, size_w, size_h, TRUE);
+   MoveWindow(wh, (int)pos_x, (int)pos_y, (int)size_w, (int)size_h, TRUE);
 #endif
 }
